ISIS-2033: house-keeping: rename some classes and packages
