--- persistence/jdo/integration/src/main/java/org/apache/isis/persistence/jdo/integration/lifecycles/_Utils.java
@@ -1,35 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package org.apache.isis.persistence.jdo.integration.lifecycles;
-
-import javax.jdo.listener.InstanceLifecycleEvent;
-
-import org.datanucleus.enhancement.Persistable;
-
-final class _Utils {
-
-    static Persistable persistableFor(InstanceLifecycleEvent event) {
-        return (Persistable)event.getSource();
-    }
-    
-    static boolean ensureRootObject(final Persistable pojo) {
-        return pojo!=null; // why would a Persistable ever be something different?
-    }
-    
-}
