+++ persistence/jdo/integration/src/main/java/org/apache/isis/persistence/jdo/integration/schema/JdoSchemaService.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.persistence.jdo.integration.session;
+package org.apache.isis.persistence.jdo.integration.schema;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
@@ -39,13 +39,19 @@ import org.apache.isis.persistence.jdo.spring.integration.TransactionAwarePersis
 
 import lombok.extern.log4j.Log4j2;
 
+/**
+ * Hooks into the application's lifecycle and runs database schema creation logic 
+ * as soon as the meta-model was populated.
+ * 
+ * @since 2.0 {@index}
+ */
 @Service
-@Named("isisJdoIntegration.JdoIntegrationService")
+@Named("isisJdoIntegration.JdoSchemaService")
 @Order(OrderPrecedence.MIDPOINT)
 @Primary
 @Qualifier("Default")
 @Log4j2
-public class JdoIntegrationService {
+public class JdoSchemaService {
 
     @Inject MetaModelContext metaModelContext;
     @Inject TransactionAwarePersistenceManagerFactoryProxy txAwarePmfProxy;
@@ -73,7 +79,7 @@ public class JdoIntegrationService {
         case PRE_METAMODEL:
             break;
         case POST_METAMODEL:
-            new DnApplication(metaModelContext, dnSettings); // creates schema
+            new _DnApplication(metaModelContext, dnSettings); // creates schema
             break;
 
         default:
